+++ project_3/include/controller_test.h
@@ -4,7 +4,7 @@
 class Controller_test
 {
 public:
-    virtual double update(double ref, double actual) = 0;
+    virtual float update(float ref, float actual) = 0;
     virtual void brake() = 0;
 };
